修改models的结构
