+++ daimaduan/forms/signin.py
@@ -3,7 +3,7 @@ from wtforms import Form
 from wtforms import StringField
 from wtforms import PasswordField
 from wtforms.validators import InputRequired
-from daimaduan.models import User
+from daimaduan.models.base import User
 
 
 class SigninForm(Form):
