+++ daimaduan/models/rate.py
@@ -1,0 +1,9 @@
+import mongoengine
+
+from daimaduan.models import BaseDocument, User, Paste
+
+
+class Rate(BaseDocument):
+    user = mongoengine.ReferenceField(User)
+    paste = mongoengine.ReferenceField(Paste)
+    score = mongoengine.IntField(default=0)
