+++ daimaduan/models/syntax.py
@@ -1,0 +1,7 @@
+import mongoengine
+
+from daimaduan.models import BaseDocument
+
+
+class Syntax(BaseDocument):
+    name = mongoengine.StringField(required=True, unique=True)
