+++ daimaduan/models/user_oauth.py
@@ -1,0 +1,11 @@
+import mongoengine
+
+from daimaduan.models import BaseDocument
+
+
+class UserOauth(BaseDocument):
+    user = mongoengine.ReferenceField('User')
+
+    provider = mongoengine.StringField(required=True)
+    openid = mongoengine.StringField(required=True)
+    token = mongoengine.StringField(required=True)
